First commit.
This is basically a copy-paste of gitbrowser, with a ton of edits of course.

It's fun writing parts of the IDE you spend time in, and this one is on the "wishlist" on Geany's site which made it a nice target.

Not fully satisified by the menu-installation trickery at the moment; will likely break for Real Users which is not exemplary.