Implemented global duplicate line removal, using a Bloom filter. Cool!
The Bloom filter is fairly straight-forward, using Austin Appleby's "MurmurHash2" from <http://sites.google.com/site/murmurhash/>.

I did try it with the current version ("MurmurHash3"), but it's way larger and more complicated, this seemed good enough.

Slightly non-traditional API for the Bloom filter; it allows you to specify a "containment" callback for exact checking, to avoid false positives. This is needed for this application, since we don't want to ever risk removing lines that have been mis-classified as dupes.

Version bumped to 2.0.